Update dir paths to reflect new locations
yjit-metrics-pages/_includes/reports is a symlink to built-yjit-reports/_includes/reports
so we can change to use it directly.

The "_benchmarks/bench_*.md" files are now in built-yjit-reports.
The "test_results" they reference are in raw-benchmark-data.

yjit-metrics-pages/raw_benchmark_data is no longer being updated and
this change removes references to it.