Disable variable warmup
The arm server has been using a config generated about 10 months ago
and been taking 9 hours to run.
If I remove the file it takes 4 hours.
If I use the latest generated file it takes over 11.

As long as the arm server isn't getting updates and we don't really
want it to run for 2 or 3 times as long,
let's also disable it for x86 to see what difference it makes.