Replace symlinks with code that specifies where to find files
Make it dynamic so that it finds ../blah on the server and ./build/blah locally.

Things should work the same way they currently do on the server (by reaching
directly for the dirs the symlinks used to target).

In local dev the site can be generated without having to do any
setup outside of the repo dir.