Ignore stats config in variable warmup to avoid abort
When the config name was fixed in 3d3db1cb7699d117139b7ed6259b5a03611aeb99
the addition of the config to the estimated time
crossed the threshold where the fabricated error is thrown.

The max time is set to 8 hours and Jenkins said it took about 4 so the
calculation must be off.

For now, since we aren't even using this report,
just ignore the stats config so jenkins can finish successfully.