Determine warmup itrs in harness: 5 for CRuby, 30 for YJIT
Stop passing it in and instead determine it at runtime.

Ideally we could take this a step further and make it adaptive
but this is a good start.

Previously the default was 10 for all configs.
If we increase warmup iterations from 10 to 30,
it should add at most 10 minutes per config
(and we run 3 YJIT configs).